backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
gitpython
google-generativeai
pypdf
python-docx
python-multipart
firebase-admin
